Splitting out Open & Update to separate file/mod
